+++ sfepy/discrete/common/extmods/refcoors.c
@@ -142,7 +142,7 @@ int32 refc_find_ref_coors_convex(FMField *ref_coors,
   int32 xi_ok, hexa_reverse;
   int32 D = mesh->topology->max_dim;
   int32 dim = D - 1;
-  int32 nc = mesh->geometry->dim;
+  uint32 nc = mesh->geometry->dim;
   uint32 tri0[] = {0, 1, 3};
   uint32 tri1[] = {2, 3, 1};
   uint32 cell, cell0, cell00, facet;
@@ -364,7 +364,7 @@ int32 refc_find_ref_coors_convex(FMField *ref_coors,
       status[ip] = 0;
     }
 
-    for (ii = 0; ii < (uint32)nc; ii++) {
+    for (ii = 0; ii < nc; ii++) {
       ref_coors->val[nc*ip+ii] = xi->val[ii];
     }
   }
